helloworld/python-client/requirements.txt
fastapi>=0.95.2
grpcio>=1.54.2
protobuf>=4.23.0,<5.0.0dev
uvicorn>=0.22.0